env file and handerr add
